MySQL을 사용해 DB 구현 완. BookList, BookDetail 구현 완.

+++ src/main/java/com/example/TeamProject/Book/BookService.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public interface BookService {
     List<BookDTO> findAll();
-    BookDTO findById(int idx);
+    BookDTO findById(long idx);
 
 }
